add function, updates
